Rename slicing test to convergence test
